fix the script short condition

+++ server/testDim.js
@@ -3,7 +3,7 @@ const startDate = '2023-06-01';
 const endDate = '2023-06-14';
 const stockId = '2330';
 const ma = [5, 10, 20]; // it can not be change, i will fix it soon.
-const type = 'long'; // long or short
+const type = 'short'; // long or short
 
 // 請勿修改下方的code
 console.log(JSON.stringify({startDate, endDate, stockId, ma, type}));
